+++ Source/Lunohod/Lunohod/Engine/XEnemy.cs
@@ -19,7 +19,7 @@ namespace Lunohod.Objects
 		/// The damage the enemy will cause when attacking the hero. Default value is 1.
 		/// </summary>
 		[XmlAttribute]
-		public float Damage = 1;
+		public double Damage = 1;
         /// <inheritdoc />
         [XmlAttribute]
         public bool IsExploding { get; set; }
